Create utils submodule; move utils.py to it
